rag-initial/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pymongo>=4.6.0

# LangChain dependencies
langchain>=0.1.0
langchain-ollama>=0.0.1
langchain-mongodb>=0.0.1
langchain-community>=0.0.1

# Document processing dependencies
PyPDF2>=3.0.0
python-docx>=1.0.0
unstructured>=0.10.0
docx2txt>=0.8

# Vector storage and embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.24.0

# LLM dependencies
llama-cpp-python>=0.3.8
torch>=2.0.0
transformers>=4.30.0

# Additional utilities
tqdm>=4.65.0
requests>=2.31.0 
